// src/interfaces/WhitelistAlphabot.ts
interface WhitelistAlphabot {
  nome: string;
  endDate: Date;
  blockchain: string;
  ogPrice: string;
  wlPrice: string;
  pubPrice: string;
  supply: string;
  qtdVencedores: string;
  slug: string;
  id: string;
  plataforma: string;
  precisaDiscordRole: boolean;
  discordRole: string;
  urlCompleta: string;
}

export { WhitelistAlphabot };